Add files via upload
this file is about create a website that can detect credit card fraud with payment gatway on webpage includes the face recognition attendance system with flask that will store data in excel format. also a software that can send otp on screen as well as on registered mobile number to prevent from online detection.